requirements.txt
boto3==1.10.1
botocore==1.13.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coremltools==3.0b3
decorator==4.4.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
graphviz==0.8.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
llvmlite==0.30.0
MarkupSafe==1.1.1
mxnet==1.5.0
numba==0.46.0
numpy==1.16.4
opencv-python==4.1.1.26
pandas==0.25.2
Pillow==6.2.1
prettytable==0.7.2
protobuf==3.10.0
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
resampy==0.2.1
s3transfer==0.2.1
scipy==1.3.1
six==1.12.0
turicreate==5.8
urllib3==1.25.6
Werkzeug==0.16.0
